LPD-25046 Move HypersonicServerClassTestRule to test/rule
