+++ modules/apps/counter/counter-test/src/testIntegration/java/com/liferay/counter/test/CounterLocalServiceProcessTest.java
@@ -7,6 +7,7 @@ package com.liferay.counter.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.counter.kernel.service.CounterLocalServiceUtil;
+import com.liferay.counter.test.rule.HypersonicServerClassTestRule;
 import com.liferay.petra.process.ProcessCallable;
 import com.liferay.petra.process.ProcessChannel;
 import com.liferay.petra.process.ProcessConfig;
