+++ modules/apps/counter/counter-test/src/testIntegration/java/com/liferay/counter/test/rule/HypersonicServerClassTestRule.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.counter.test;
+package com.liferay.counter.test.rule;
 
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.CharPool;
